+++ src/gui/snuffler.py
@@ -17,7 +17,7 @@ from os.path import join as pjoin
 from optparse import OptionParser
 
 
-from pyrocko import pile
+from pyrocko import pile as pile_mod
 from pyrocko import util
 from pyrocko import model
 from pyrocko import config
@@ -74,12 +74,14 @@ def snuffle(pile=None, **kwargs):
     :param want_markers: bool, whether markers should be returned
     :param launch_hook: callback function called before snuffler window is
         shown
+    :param instant_close: bool, whether to bypass close window confirmation
+        dialog
     '''
     from .snuffler_app import SnufflerWindow, \
         setup_acquisition_sources, PollInjector
 
     if pile is None:
-        pile = pile.make_pile()
+        pile = pile_mod.make_pile()
 
     app = get_snuffler_instance()
 
@@ -298,7 +300,7 @@ def snuffler_from_commandline(args=None):
     if options.gui_toolkit_qt5:
         config.override_gui_toolkit = 'qt5'
 
-    this_pile = pile.Pile()
+    this_pile = pile_mod.Pile()
     stations = []
     for stations_fn in extend_paths(options.station_fns):
         stations.extend(model.station.load_stations(stations_fn))