+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   problems when included in XeLaTeX).
 - QuakeML: Not strictly requiring preferred origin to be set anymore when
   extracting Pyrocko event objects.
+- Snuffler now asks for confirmation when the user attempts to close the
+  window.
 
 ## [2020.10.26]